Create binary file

+++ CPU/config_cmd.h
@@ -1,5 +1,5 @@
 
-const int CMD_PUSH = 1;
-const int CMD_ADD  = 2;
-const int CMD_OUT  = 3;
-const int CMD_HLT  = 4;
+const size_t CMD_PUSH = 1;
+const size_t CMD_ADD  = 2;
+const size_t CMD_OUT  = 3;
+const size_t CMD_HLT  = 4;
